+++ ExtraSlots.cs
@@ -395,16 +395,13 @@ namespace ExtraSlots
             extraRows.SettingChanged += (s, e) => API.UpdateSlots();
             rowsProgressionEnabled.SettingChanged += (s, e) => API.UpdateSlots();
             
-            extraUtilitySlotsAmount.SettingChanged += (s, e) =>
-            {
-                EquipmentPanel.UpdatePanel();
-                Compatibility.EpicLootCompat.InvalidatePlayerEffectCache(Player.m_localPlayer);
-            };
-            quickSlotsAmount.SettingChanged += (s, e) => EquipmentPanel.UpdatePanel();
-            foodSlotsEnabled.SettingChanged += (s, e) => EquipmentPanel.UpdatePanel();
-            miscSlotsEnabled.SettingChanged += (s, e) => EquipmentPanel.UpdatePanel();
-            ammoSlotsEnabled.SettingChanged += (s, e) => EquipmentPanel.UpdatePanel();
-            slotsProgressionEnabled.SettingChanged += (s, e) => EquipmentPanel.UpdatePanel();
+            extraUtilitySlotsAmount.SettingChanged += (s, e) => API.UpdateSlotActivation();
+            quickSlotsAmount.SettingChanged += (s, e) => API.UpdateSlotActivation();
+            foodSlotsEnabled.SettingChanged += (s, e) => API.UpdateSlotActivation();
+            miscSlotsEnabled.SettingChanged += (s, e) => API.UpdateSlotActivation();
+            ammoSlotsEnabled.SettingChanged += (s, e) => API.UpdateSlotActivation();
+            slotsProgressionEnabled.SettingChanged += (s, e) => API.UpdateSlotActivation();
+            customSlotItemsCanUseRegularEquipmentSlots.SettingChanged += (s, e) => API.UpdateSlotActivation();
             preventUniqueUtilityItemsEquip.SettingChanged += (s, e) => ExtraUtilitySlots.UpdateUniqueEquipped();
             showExtraUtilityItems.SettingChanged += (s,e) => Player.m_localPlayer?.SetupEquipment();
 
@@ -533,8 +530,16 @@ namespace ExtraSlots
                     GetDescriptionSeparatedStrings("Comma separated list of items that should NOT be treated as misc items" +
                                             "\nWorks with prefab names (like BeltStrength) and item names (like $item_beltstrength). [Synced with Server]"));
 
-            miscSlotsItemList.SettingChanged += (s, e) => Slots.UpdateMiscSlotCustomItemList();
-            miscSlotsItemBlackList.SettingChanged += (s, e) => Slots.UpdateMiscSlotCustomItemList();
+            miscSlotsItemList.SettingChanged += (s, e) =>
+            {
+                Slots.UpdateMiscSlotCustomItemList();
+                API.UpdateSlotActivation();
+            };
+            miscSlotsItemBlackList.SettingChanged += (s, e) =>
+            {
+                Slots.UpdateMiscSlotCustomItemList();
+                API.UpdateSlotActivation();
+            };
 
             quickSlotsHotBarEnabled = serverConfig("Panels - Quick slots", "Enabled", defaultValue: true, "Enable hotbar with quick slots [Synced with Server]");
             quickSlotsHotBarOffset = config("Panels - Quick slots", "Offset", defaultValue: new Vector2(230f, 156f), "On screen position of quick slots hotbar panel");
@@ -578,8 +583,16 @@ namespace ExtraSlots
                     GetDescriptionSeparatedStrings("Comma separated list of items that should NOT be treated as ammo items to fit in ammo slots" +
                                             "\nWorks with prefab names (like BeltStrength) and item names (like $item_beltstrength). [Synced with Server]"));
 
-            ammoSlotsItemList.SettingChanged += (s, e) => Slots.UpdateAmmoSlotCustomItemList();
-            ammoSlotsItemBlackList.SettingChanged += (s, e) => Slots.UpdateAmmoSlotCustomItemList();
+            ammoSlotsItemList.SettingChanged += (s, e) =>
+            {
+                Slots.UpdateAmmoSlotCustomItemList();
+                API.UpdateSlotActivation();
+            };
+            ammoSlotsItemBlackList.SettingChanged += (s, e) =>
+            {
+                Slots.UpdateAmmoSlotCustomItemList();
+                API.UpdateSlotActivation();
+            };
 
             ammoSlotsHotBarEnabled.SettingChanged += (s, e) => HotBars.AmmoSlotsHotBar.MarkDirty();
             ammoSlotsHotBarOffset.SettingChanged += (s, e) => HotBars.AmmoSlotsHotBar.MarkDirty();
@@ -607,8 +620,16 @@ namespace ExtraSlots
                     GetDescriptionSeparatedStrings("Comma separated list of items that should NOT be treated as ammo items to fit in food slots" +
                                             "\nWorks with prefab names (like BeltStrength) and item names (like $item_beltstrength). [Synced with Server]"));
 
-            foodSlotsItemList.SettingChanged += (s, e) => Slots.UpdateFoodSlotCustomItemList();
-            foodSlotsItemBlackList.SettingChanged += (s, e) => Slots.UpdateFoodSlotCustomItemList();
+            foodSlotsItemList.SettingChanged += (s, e) =>
+            {
+                Slots.UpdateFoodSlotCustomItemList();
+                API.UpdateSlotActivation();
+            };
+            foodSlotsItemBlackList.SettingChanged += (s, e) =>
+            {
+                Slots.UpdateFoodSlotCustomItemList();
+                API.UpdateSlotActivation();
+            };
 
             foodSlotsHotBarEnabled.SettingChanged += (s, e) => HotBars.FoodSlotsHotBar.MarkDirty();
             foodSlotsHotBarOffset.SettingChanged += (s, e) => HotBars.FoodSlotsHotBar.MarkDirty();
@@ -672,36 +693,36 @@ namespace ExtraSlots
             foodSlotsAvailableAfterDiscovery = serverConfig("Progression - Discovery", "Food slots", true, "Food slots will be active after acquiring first food item [Synced with Server]");
             equipmentSlotsAvailableAfterDiscovery = serverConfig("Progression - Discovery", "Equipment slots", true, "Corresponding equipment slot will be active after acquiring first item [Synced with Server]");
 
-            quickSlotGlobalKey1.SettingChanged += (s, e) => EquipmentPanel.UpdatePanel();
-            quickSlotGlobalKey2.SettingChanged += (s, e) => EquipmentPanel.UpdatePanel();
-            quickSlotGlobalKey3.SettingChanged += (s, e) => EquipmentPanel.UpdatePanel();
-            quickSlotGlobalKey4.SettingChanged += (s, e) => EquipmentPanel.UpdatePanel();
-            quickSlotGlobalKey5.SettingChanged += (s, e) => EquipmentPanel.UpdatePanel();
-            quickSlotGlobalKey6.SettingChanged += (s, e) => EquipmentPanel.UpdatePanel();
+            quickSlotGlobalKey1.SettingChanged += (s, e) => API.UpdateSlotActivation();
+            quickSlotGlobalKey2.SettingChanged += (s, e) => API.UpdateSlotActivation();
+            quickSlotGlobalKey3.SettingChanged += (s, e) => API.UpdateSlotActivation();
+            quickSlotGlobalKey4.SettingChanged += (s, e) => API.UpdateSlotActivation();
+            quickSlotGlobalKey5.SettingChanged += (s, e) => API.UpdateSlotActivation();
+            quickSlotGlobalKey6.SettingChanged += (s, e) => API.UpdateSlotActivation();
 
-            ammoSlotsGlobalKey.SettingChanged += (s, e) => EquipmentPanel.UpdatePanel();
-            foodSlotsGlobalKey.SettingChanged += (s, e) => EquipmentPanel.UpdatePanel();
-            miscSlotsGlobalKey.SettingChanged += (s, e) => EquipmentPanel.UpdatePanel();
+            ammoSlotsGlobalKey.SettingChanged += (s, e) => API.UpdateSlotActivation();
+            foodSlotsGlobalKey.SettingChanged += (s, e) => API.UpdateSlotActivation();
+            miscSlotsGlobalKey.SettingChanged += (s, e) => API.UpdateSlotActivation();
 
-            utilitySlotGlobalKey1.SettingChanged += (s, e) => EquipmentPanel.UpdatePanel();
-            utilitySlotGlobalKey2.SettingChanged += (s, e) => EquipmentPanel.UpdatePanel();
-            utilitySlotGlobalKey3.SettingChanged += (s, e) => EquipmentPanel.UpdatePanel();
-            utilitySlotGlobalKey4.SettingChanged += (s, e) => EquipmentPanel.UpdatePanel();
+            utilitySlotGlobalKey1.SettingChanged += (s, e) => API.UpdateSlotActivation();
+            utilitySlotGlobalKey2.SettingChanged += (s, e) => API.UpdateSlotActivation();
+            utilitySlotGlobalKey3.SettingChanged += (s, e) => API.UpdateSlotActivation();
+            utilitySlotGlobalKey4.SettingChanged += (s, e) => API.UpdateSlotActivation();
 
-            ammoSlotsAvailableAfterDiscovery.SettingChanged += (s, e) => EquipmentPanel.UpdatePanel();
-            utilitySlotAvailableAfterDiscovery.SettingChanged += (s, e) => EquipmentPanel.UpdatePanel();
-            foodSlotsAvailableAfterDiscovery.SettingChanged += (s, e) => EquipmentPanel.UpdatePanel();
-            equipmentSlotsAvailableAfterDiscovery.SettingChanged += (s, e) => EquipmentPanel.UpdatePanel();
+            ammoSlotsAvailableAfterDiscovery.SettingChanged += (s, e) => API.UpdateSlotActivation();
+            utilitySlotAvailableAfterDiscovery.SettingChanged += (s, e) => API.UpdateSlotActivation();
+            foodSlotsAvailableAfterDiscovery.SettingChanged += (s, e) => API.UpdateSlotActivation();
+            equipmentSlotsAvailableAfterDiscovery.SettingChanged += (s, e) => API.UpdateSlotActivation();
 
             utilitySlotItemDiscovered1 = serverConfig("Progression - Items", "Extra utility slot 1", "Wishbone", "Comma-separated list of items. Slot will be active only if any item is discovered or list is not set. [Synced with Server]");
             utilitySlotItemDiscovered2 = serverConfig("Progression - Items", "Extra utility slot 2", "Demister,GoldRubyRing,SilverRing", "Comma-separated list of items. Slot will be active only if any item is discovered or list is not set. [Synced with Server]");
             utilitySlotItemDiscovered3 = serverConfig("Progression - Items", "Extra utility slot 3", "", "Comma-separated list of items. Slot will be active only if any item is discovered or list is not set. [Synced with Server]");
             utilitySlotItemDiscovered4 = serverConfig("Progression - Items", "Extra utility slot 4", "", "Comma-separated list of items. Slot will be active only if any item is discovered or list is not set. [Synced with Server]");
 
-            utilitySlotItemDiscovered1.SettingChanged += (s, e) => EquipmentPanel.UpdatePanel();
-            utilitySlotItemDiscovered2.SettingChanged += (s, e) => EquipmentPanel.UpdatePanel();
-            utilitySlotItemDiscovered3.SettingChanged += (s, e) => EquipmentPanel.UpdatePanel();
-            utilitySlotItemDiscovered4.SettingChanged += (s, e) => EquipmentPanel.UpdatePanel();
+            utilitySlotItemDiscovered1.SettingChanged += (s, e) => API.UpdateSlotActivation();
+            utilitySlotItemDiscovered2.SettingChanged += (s, e) => API.UpdateSlotActivation();
+            utilitySlotItemDiscovered3.SettingChanged += (s, e) => API.UpdateSlotActivation();
+            utilitySlotItemDiscovered4.SettingChanged += (s, e) => API.UpdateSlotActivation();
 
             quickSlotItemDiscovered1 = serverConfig("Progression - Items", "Quickslot 1", "CryptKey", "Comma-separated list of items. Slot will be active only if any item is discovered or list is not set. [Synced with Server]");
             quickSlotItemDiscovered2 = serverConfig("Progression - Items", "Quickslot 2", "CryptKey", "Comma-separated list of items. Slot will be active only if any item is discovered or list is not set. [Synced with Server]");
@@ -710,12 +731,12 @@ namespace ExtraSlots
             quickSlotItemDiscovered5 = serverConfig("Progression - Items", "Quickslot 5", "", "Comma-separated list of items. Slot will be active only if any item is discovered or list is not set. [Synced with Server]");
             quickSlotItemDiscovered6 = serverConfig("Progression - Items", "Quickslot 6", "", "Comma-separated list of items. Slot will be active only if any item is discovered or list is not set. [Synced with Server]");
 
-            quickSlotItemDiscovered1.SettingChanged += (s, e) => EquipmentPanel.UpdatePanel();
-            quickSlotItemDiscovered2.SettingChanged += (s, e) => EquipmentPanel.UpdatePanel();
-            quickSlotItemDiscovered3.SettingChanged += (s, e) => EquipmentPanel.UpdatePanel();
-            quickSlotItemDiscovered4.SettingChanged += (s, e) => EquipmentPanel.UpdatePanel();
-            quickSlotItemDiscovered5.SettingChanged += (s, e) => EquipmentPanel.UpdatePanel();
-            quickSlotItemDiscovered6.SettingChanged += (s, e) => EquipmentPanel.UpdatePanel();
+            quickSlotItemDiscovered1.SettingChanged += (s, e) => API.UpdateSlotActivation();
+            quickSlotItemDiscovered2.SettingChanged += (s, e) => API.UpdateSlotActivation();
+            quickSlotItemDiscovered3.SettingChanged += (s, e) => API.UpdateSlotActivation();
+            quickSlotItemDiscovered4.SettingChanged += (s, e) => API.UpdateSlotActivation();
+            quickSlotItemDiscovered5.SettingChanged += (s, e) => API.UpdateSlotActivation();
+            quickSlotItemDiscovered6.SettingChanged += (s, e) => API.UpdateSlotActivation();
 
             extraRowPlayerKey1 = serverConfig("Progression - Inventory - Player keys", "Extra row 1", "GP_Bonemass", "Comma-separated list of Player unique keys. Extra inventory row will be active only if any key is enabled or list is not set. [Synced with Server]");
             extraRowPlayerKey2 = serverConfig("Progression - Inventory - Player keys", "Extra row 2", "GP_Moder", "Comma-separated list of Player unique keys. Extra inventory row will be active only if any key is enabled or list is not set. [Synced with Server]");